requirements.txt
blinker==1.9.0
cachetools==5.5.2
click==8.2.1
colorama==0.4.6
Flask
flask-cors
python-dotenv
Werkzeug
google-auth
google-auth-oauthlib
mysql-connector-python
rsa==4.9.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1